Smoothed things out

+++ includedFiles/demo4_5/runExec.sh
@@ -9,28 +9,42 @@ MEASURE_PORT=3057
 APPLICATION=test1.o
 
 DEBUG=1
+echo "" > ~/log.1
 
+
+rm ~/screenlog.0
 if [ "$EXEC" == "Attester" ];then
-   echo "HERE"
    #Check to see if tpmd is running
 
    tpmd_running=`ps x| grep -v "grep" | grep tpmd`
 
+   #start tpmd if not running
    if [ "$tpmd_running" == "" ]; then
        tpmd
    fi
 
-   killall $APPLICATION
-   killall gdb
+   #kill Application and gdb if already running
+   if [ ! -z "`pidof $APPLICATION`" ];then
+      killall $APPLICATION
+      sleep 1
+   fi 
+   
+   if [ ! -z "`pidof gdb`" ];then
+       killall gdb
+       sleep 1
+   fi
 
-   screen -dmS app_session $DIR/$APPLICATION
+
+  #start Application
+   screen -dmS app_session $DIR/$APPLICATION 
    PID=`pidof $APPLICATION`
-   echo "Application: $PID"
+   echo "$APPLICATION PID: $PID" >> ~/log.1
 
-   screen -dmS measurer_screen $DIR/gdb --port=$MEASURE_PORT > measurer_out
+   #measurer output goes to screenlog.0
+   screen -dmLS measurer_screen $DIR/gdb --port=$MEASURE_PORT
 
    if [ "$PID" == "" ];then
-       echo "Error in launching $APPLICATION"
+       echo "Error in launching $APPLICATION try again" >> ~/log.1
        exit
    fi
 
@@ -43,8 +57,9 @@ else
 fi
 
 if [ "$PID" != "" ];then
-   $DIR/$EXEC  $APP $ATT $CA $MEASURE_PORT $PID $DEBUG &
+   $DIR/$EXEC  $APP $ATT $CA $MEASURE_PORT $PID $DEBUG 2>> ~/log.1 &
 else
    $DIR/$EXEC  $APP $ATT $CA &
 fi
 
+echo "Successfull launch" >> ~/log.1
